+++ Source/Scenarios/RealisticScenario.hpp
@@ -36,17 +36,17 @@ namespace Scenarios {
       double   dy;
     };
 
-    bool loadBinaryData(const std::string& filename, FileHeader& header, Float2D<double>& data);
+    bool loadBinaryData(const std::string& filename, FileHeader& header, Float2D<float>& data);
 
     BoundaryType boundaryType_;
 
-    Float2D<double> b_;
+    Float2D<float> b_;
     int               bNX_, bNY_;
     double            boundaryPos_[4];
     RealType          originX_, originY_;
     RealType          bDX_, bDY_;
 
-    Float2D<double> d_;
+    Float2D<float> d_;
     int               dNX_, dNY_;
     double            dBoundaryPos_[4];
     RealType          dOriginX_, dOriginY_;
